pyproject.toml
[tool.poetry]
name = "IssuesTgBot"
version = "0.1.0"
description = ""
authors = ["Mark Minakov <maminakov@vk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.7"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

